Rename backend source files.
If we want to be able to compile more than one backend, it will be good to have
source files named differently in order to reduce confusion.

This change DOES NOT COMPILE. The reason it was committed separately was to allo git
to properly track renames.
