+++ simpleble/src/backends/linux/AdapterLinux.h
