+++ simpleble/src/backends/plain/AdapterPlain.cpp
@@ -1,4 +1,4 @@
-#include "AdapterBase.h"
+#include "AdapterPlain.h"
 #include "CommonUtils.h"
 #include "PeripheralBase.h"
 #include "PeripheralBuilder.h"
