+++ simpleble/src/backends/plain/AdapterPlain.h
